BF-50: moved articles ui to its own folder + wip articles screen
