+++ lib/models/article_image.dart
@@ -1,0 +1,7 @@
+class ArticleImage {
+  final String url;
+  final double width;
+  final double height;
+
+  ArticleImage({this.url, this.width, this.height});
+}
